remove justtee from most scripts, it hides test failures in full build
By piping output to justtee (to write mocha and jshint output to
a file (for project report) as well as to the console, if linting
or tests fail, the exit code of the whole npm script step will still
be 0 because the exit code of justtee is always 0.

This makes the build pass even if there are test errors.

We still need a solution for writing the project reports, though,
but that has a lower priority.